Move AtFwd hal from device to framework manifest.
AtFwd hal is implemented as an android service and
it starts from system partition, that is, it's not
a traditional vendor hal and hence it should be
declared in the framework manifest and not in the
device manifest.
This commit introduces new device specific framework
manifest (framework_manifest.xml) and moves AtFwd hal
from device to framework manifest.

Change-Id: I8c1d3a4f3bade6fbdfa0211ebf2424440a0b4970
CRs-Fixed: 2311156