init.target.rc: set usb properties on charger
In charging mode, USB may be connected to a PC port. In order
to charge the battery at up to 500mA (or 900mA in the case of
SuperSpeed), the USB configuration needs to enumerate. For the
composition to apply, sys.usb.configfs property set to 1 and
the correct controller device needs to be specified similar
to normal boot. Add these properties to 'on charger' trigger
to allow the correct USB instance to be started.

Change-Id: I031ed3ddec629da6589ab6115027d259b188a37a